Fixed problem with events not loading if month before that event had no events and also fixed lots of things and stuff that I can't remember

+++ campuslife/CalendarViewController.m
@@ -124,6 +124,7 @@
     [self.navigationItem setHidesBackButton:YES animated:YES];
     
     [self getEventsForMonth:[_events getSelectedMonth] :[_events getSelectedYear]];
+
     
 
 }
@@ -1555,7 +1556,7 @@
         [self getEventsForMonth:[_events getSelectedMonth] :[_events getSelectedYear]];
         
         
-        [_events setSelectedDay:day];
+        [_events setSelectedDay:(int)day];
         [self.collectionView reloadData];
     }
 }
